+++ front/src/contexts/fieldContext.tsx
@@ -2,6 +2,7 @@ import { createContext, useContext } from 'react';
 
 const FIELD_CONTEXT_DEFAULT_VALUE = {
   isValid: true,
+  htmlFor: '',
 };
 export const fieldContext = createContext(FIELD_CONTEXT_DEFAULT_VALUE);
 export const useFieldContext = () => useContext(fieldContext);
